+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Update: Golang and npm packages/security patches.
+- Update: If filters are applied to the table, the CSV export will only export the filtered rows.
 
 ## [4.0.1] - 08.03.2026